date 27.05.2020

+++ database/src/main/java/by/itacademy/boldysh/database/dto/CarDto.java
@@ -1,12 +1,13 @@
 package by.itacademy.boldysh.database.dto;
 
-import by.itacademy.boldysh.database.entity.BrandCar;
 import by.itacademy.boldysh.database.entity.CarClass;
 import by.itacademy.boldysh.database.entity.Transmission;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 
 
 @Data
@@ -19,7 +20,7 @@ public class CarDto {
     private String brandCar;
     private String vinNumber;
     private Integer yearOfIssue;
-    private Integer costRentalOfDay;
+    private BigDecimal costRentalOfDay;
     private Transmission transmission;
     private CarClass classCar;
