+++ service/src/main/java/by/itacademy/boldysh/service/interfaces/AdditionalServiceService.java
@@ -4,8 +4,10 @@ import by.itacademy.boldysh.database.entity.AdditionalService;
 import by.itacademy.boldysh.database.entity.Services;
 import by.itacademy.boldysh.service.util.BaseMethod;
 
+import java.math.BigDecimal;
+
 public interface AdditionalServiceService extends BaseMethod<Long, AdditionalService> {
 
-    void updateAdditionalService(Services additionalService, Integer cost);
+    void updateAdditionalService(Services additionalService, BigDecimal cost);
 
 }
